userAuth and user-associated boards validations completed

+++ src/controllers/board.controllers.js
@@ -13,8 +13,9 @@ boardCtrl.createNewBoard = async (req,res) => {
     const boardTitle = req.body.title;
     const boardDescription = req.body.description;
     const newBoard = new board({boardTitle,boardDescription});
-    console.log(newBoard);
+    newBoard.createdBy = req.user.id;
     await newBoard.save();
+
     req.flash('success', 'Board created succesfully');
     res.redirect('/board');
 };
@@ -29,7 +30,7 @@ boardCtrl.createNewCard = (req,res) => {
 };
 
 boardCtrl.renderBoards = async (req,res) => {
-    const boards = await board.find();
+    const boards = await board.find({createdBy: req.user.id}).sort({createdAt: 'desc'});
     res.render('./board/board', {boards}); // carpeta board, archivo board.ejs
 };
 
@@ -39,12 +40,20 @@ boardCtrl.renderCards = (req,res) => {
 };
 
 boardCtrl.renderEditForm = async (req,res) => {
-    const currentBoard = await (await board.findById(req.params.id));
-    console.log(currentBoard);
+    const currentBoard = await board.findById(req.params.id);
+    if (currentBoard.createdBy != req.user.id) {
+        req.flash('error', 'You can only edit your boards');
+        return res.redirect('/board');
+    }
     res.render('board/boardedit', {currentBoard});
 };
 
 boardCtrl.deleteBoard = async (req,res) => {
+    const currentBoard = await board.findById(req.params.id);
+    if (currentBoard.createdBy != req.user.id) { // Unnecesary because this works on a POST method
+        req.flash('error', 'You can only delete your boards');
+        return res.redirect('/board');
+    }
     await board.findByIdAndDelete(req.params.id)
     req.flash('success', 'Your board was deleted');
     res.redirect('/board');
